% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adf.R
\name{adf}
\alias{adf}
\title{Type less words for as.data.frame}
\usage{
adf(x)
}
\arguments{
\item{x}{a object to transform to a data.frame()}
}
\value{
a object of class data.frame
}
\description{
Type less words for as.data.frame
}
\examples{
class(adf(Orange$age))
}
