% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_na.R
\name{all_na}
\alias{all_na}
\title{Determine all NA columns}
\usage{
all_na(x)
}
\arguments{
\item{x}{}
}
\value{
a string indicating the numeric column position where there
   are all NAs
}
\description{
Determine all NA columns
}
\seealso{
\url{http://stackoverflow.com/a/11330265/4143444}
}
