% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_smooth_mean.R
\name{bin_smooth_mean}
\alias{bin_smooth_mean}
\title{Bin smoothing by mean}
\usage{
bin_smooth_mean(x, bins)
}
\arguments{
\item{x}{a numeric vector or list of numbers}

\item{bins}{a numeric vector of how many bins}
}
\value{
a list
}
\description{
Bin smoothing by mean
}
\examples{
bin_smooth_mean(x = c(21, 15, 26, 26, 28, 29), bins=2)
}

