% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_smooth_median.R
\name{bin_smooth_median}
\alias{bin_smooth_median}
\title{Bin smoothing by median}
\usage{
bin_smooth_median(x, bins)
}
\arguments{
\item{x}{a numeric vector or list of numbers}

\item{bins}{a numeric vector of how many bins}
}
\value{
a list
}
\description{
Bin smoothing by median
}
\examples{
bin_smooth_median(x = c(21, 15, 26, 26, 28, 29), bins=2)
}
