% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{birthday}
\alias{birthday}
\title{A simple function for Birthday awareness}
\usage{
birthday()
}
\value{
a printed message indicating if it is my birthday
}
\description{
A simple function for Birthday awareness
}
\examples{
 birthday()
}
