% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind.fill.R
\name{cbind.fill}
\alias{cbind.fill}
\title{cbind.fill}
\usage{
cbind.fill(...)
}
\arguments{
\item{...}{data.frame objects needed to be combined by similar columns}
}
\value{
a data.frame
}
\description{
returns a data.frame joined by columns but filled with NA's if the values are missing
}
\examples{
x = data.frame("x1" = c(1, 2, 3))
y = data.frame("x1" = c(1, 2), "y1" = c(1, 2))
cbind.fill(x, y)
}
\seealso{
http://stackoverflow.com/questions/7962267/cbind-a-df-with-an-empty-df-cbind-fill
}

