% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ge.R
\name{clean_ge}
\alias{clean_ge}
\title{Clean Global Environment
Remove everything in your Global Environment except the function that does it}
\usage{
clean_ge(pattern = "clean_ge")
}
\arguments{
\item{pattern}{the name of the function}
}
\value{
nothing
}
\description{
Clean Global Environment
Remove everything in your Global Environment except the function that does it
}
\examples{
# clean_ge()
}
\seealso{
url{http://stackoverflow.com/questions/4837477/remove-objects-in-globalenv-from-within-a-function}
}
