% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_block.R
\name{code_block}
\alias{code_block}
\title{add a code block to your code!}
\usage{
code_block(text)
}
\arguments{
\item{text}{a character string}
}
\value{
formated string in the console to be copied to your source
}
\description{
add a code block to your code!
}
\examples{
code_block("This section is important")

## output is below to be copied:
# # # # # # # # # # # # # # # # # # # # # # # # #
  #  This section is important
# # # # # # # # # # # # # # # # # # # # # # # # #
}
