% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_chr_to_factor.R
\name{convert_chr_to_factor}
\alias{convert_chr_to_factor}
\title{convert character strings into factors}
\usage{
convert_chr_to_factor(df, ...)
}
\arguments{
\item{...}{dataframe with characters}
}
\value{
df
}
\description{
convert character strings into factors
}
\details{
http://stackoverflow.com/a/2851213
}

