% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_file_path.R
\name{create_file_path}
\alias{create_file_path}
\title{Create a file path}
\usage{
create_file_path(mainDir, subDir, version_num, ...)
}
\arguments{
\item{mainDir}{a object of class 'character' which indicates your root that is already created}

\item{subDir}{a object of class 'character' which indicates your sub directory for models}

\item{version_num}{a object of class character which indicates your sub directory for model iterations}

\item{...}{}
}
\value{
a file folder is created if logic is satisfied
}
\description{
Create a file path
}
