% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.test.R
\name{cv.test}
\alias{cv.test}
\title{Title}
\usage{
cv.test(x, y, ...)
}
\arguments{
\item{x}{a numeric vector or matrix, or list to be compared against y}

\item{y}{a numeric vector or matrix, or list to be compared against x}

\item{...}{}
}
\value{
A list with class "htest"
}
\description{
Title
}
\details{
http://www.r-bloggers.com/example-8-39-calculating-cramers-v/
}

