% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_test.R
\name{cv_test}
\alias{cv_test}
\title{calculate the cramer's V statistic}
\usage{
cv_test(x, y, ...)
}
\arguments{
\item{x}{a numeric vector or matrix, or list to be compared against y}

\item{y}{a numeric vector or matrix, or list to be compared against x}

\item{...}{optional arguements}
}
\value{
A list with class "htest"
}
\description{
calculate the cramer's V statistic
}
\seealso{
\url{http://www.r-bloggers.com/example-8-39-calculating-cramers-v/}
}

