% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_cramersV_hand.R
\name{example_cramersV_hand}
\alias{example_cramersV_hand}
\title{example Cramer's V function}
\usage{
example_cramersV_hand()
}
\description{
computes the Cramer's V statisitc with dummy data using the lsr funciton and a hand equation
}
\examples{
example_cv_func()
}

