% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Opposite of %in%}
\usage{
x \%notin\% table
}
\arguments{
\item{x}{left match value comparison}

\item{table}{right match table comparison}
}
\value{
the values not in the table
}
\description{
Opposite of %in%
}
\examples{
"xc5" \%notin\% letters
"c" \%notin\% letters
}
