% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has.negative.R
\name{has.negative}
\alias{has.negative}
\title{find negative values in a given data frame}
\usage{
has.negative(df, ...)
}
\arguments{
\item{df}{an object of class data.frame}

\item{...}{}
}
\description{
find negative values in a given data frame
}
\details{
this function is in development as it returns multiple statements if a negative value is found
}

