% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello!}
\usage{
hello()
}
\arguments{
\item{""}{no arguement needed}
}
\value{
"Hi, I'm Jasmine"
}
\description{
This is a starter function to just say hello!
}
\examples{
hello()
}
