% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_xgmat.R
\name{make_xgmat}
\alias{make_xgmat}
\title{Create a Sparse matrix for xgboost}
\usage{
make_xgmat(data, label, ...)
}
\arguments{
\item{data}{an object of class data.frame or tbl.df}

\item{...}{optional arguements}

\item{train_label}{a object of class numeric}

\item{test_label}{a object of class numeric}

\item{type}{a object of class string}
}
\value{
return a xgb.Matrix to the Global environment
}
\description{
Create a Sparse matrix for xgboost
}
\examples{
make_xgmat(traindat, train_label)
}

