% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_to_zero.R
\name{missing_to_zero}
\alias{missing_to_zero}
\title{Transform Missing NA values to Zeros}
\source{
\url{http://stackoverflow.com/questions/2641653/pass-a-data-frame-column-name-to-a-function}
}
\usage{
missing_to_zero(df, ...)
}
\arguments{
\item{df}{a data frame or indicated column of the data frame (named list)}

\item{...}{a character vector optionally supplied as a c(...) vector}
}
\value{
returns the column of interest with transformed missing NA values to 0
}
\description{
Transform NA values to Zeros, which can be helpful
 in data cleaning for loss amounts
}
\examples{
df = data.frame("x" = 1:3, "missings" = rep(NA, 3))
df$missings <- missing_to_zero(df, "missings")
}

