% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_levels.R
\name{multi_levels}
\alias{multi_levels}
\title{Multi-levels in a data.frame
Determine how many levels are in a data.frame}
\usage{
multi_levels(df, ...)
}
\arguments{
\item{df}{a object of class data.frame}

\item{...}{optional parameters supplied}
}
\value{
an object of class 'list' indicating the column name and the amount of levels
}
\description{
Multi-levels in a data.frame
Determine how many levels are in a data.frame
}
\examples{
sample_df = data.frame("x" = rnorm(1:10),
                    "y" = rnorm(1:10),
                    "letters" = letters[1:10],
                    "gender" = c("M", "F"),
                    "pet" = c("spike", "spot", "fluffy", "mr. jingles", "rex") )

multi_levels(df=sample_df)
}
