% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.this.R
\name{open.this}
\alias{open.this}
\title{Title}
\usage{
\method{open}{this}(x, file = "__tmp.csv")
}
\arguments{
\item{x}{dataset or object of class data.frame}

\item{file}{default file append}
}
\value{
"press enter to continue..."
}
\description{
This opens a dataset in a tempfile in excel for cleaner visuals of the data
}
\examples{
open.this()
}
\seealso{
Function graciously acquired from Garrett Dancik, PhD @ ECSU
}

