% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelDiv.R
\name{panelDiv}
\alias{panelDiv}
\title{Add a panel div block for Shiny}
\usage{
panelDiv(class_type, panel_title, content)
}
\arguments{
\item{class_type}{object of class character, options for color 'info', 'primary', 'success', 'danger', 'warning'}

\item{panel_title}{object of class character, text for panel header.}

\item{content}{object of class character, functions for date or time}
}
\value{
HTML formated for a panel div
}
\description{
Add a panel div block for Shiny
}
\details{
http://getbootstrap.com/components/#panels
}
\examples{
panelDiv('info', 'Status', Sys.Date())
}

