% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{percent}
\alias{percent}
\title{calculate a percentage}
\usage{
percent(df)
}
\arguments{
\item{df}{an object of class data.frame or list of numeric or double values}
}
\value{
"numeric"
}
\description{
calculate the percent of a list of numerical objects in R
}
\examples{
df = c(1, 4, 6, 7, 3, 12)
percent(df)
}

