% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performr.R
\name{performr}
\alias{performr}
\title{Test and compare csv reading performance with multiple packages}
\usage{
performr(filename, method = "all", ...)
}
\arguments{
\item{filename}{object of class character indicating filepath of dataset}

\item{method}{object of class character/string choose "all" to run all the methods or select from single methods 'readr','utils', 'data.table', 'bigmemory', 'ff', 'sqldf'.}

\item{...}{optional}
}
\value{
a string message stating which method was the fastest
}
\description{
Test and compare csv reading performance with multiple packages
}
\examples{
## don't run
## performr("iris.csv", "readr")
}
