% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_only_factors.R
\name{replace_only_factors}
\alias{replace_only_factors}
\title{Replace only the factors in a dataframe}
\usage{
replace_only_factors(df)
}
\arguments{
\item{df}{object of class data.frame}
}
\value{
returns the dataframe with characters in place of factors
}
\description{
Replace only the factors in a dataframe
}
\details{
http://stackoverflow.com/a/2853231
}

