% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_plot.R
\name{roc_plot}
\alias{roc_plot}
\title{ROC Curve for AUC Metric
Produce a ROC curve}
\usage{
roc_plot(model, test_data, test_label, ...)
}
\arguments{
\item{model}{a object of class model (glm, lm, etc...)}

\item{test_data}{a object for the test data}

\item{test_label}{a object for the test label}
}
\value{
a data visualization depicting the AUC and baseline
}
\description{
ROC Curve for AUC Metric
Produce a ROC curve
}

