% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_check.R
\name{sanity_check}
\alias{sanity_check}
\title{Check your sanity and the length or rows for equality}
\usage{
sanity_check(check1, check2, ...)
}
\arguments{
\item{check1}{a object of class data.frame, character, or numeric}

\item{check2}{a object of class data.frame, character, or numeric}

\item{...}{optional arguements passed through (TBD)}
}
\value{
a string indicating the success of the comparison or
 failure of equality of observations form both input objects
}
\description{
After creating train and test datasets in ML, check to make
 sure that the correponsing weights, or response vectors are the same
 amount of observations as the original dataset.
}
\examples{
sanity_check(iris, iris)
sanity_check(iris, cars)
}
