% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_mail.R
\name{send_email}
\alias{send_email}
\title{Send a email upon R code completion}
\usage{
send_email(subject, text, address, file = FALSE, filename = "MyRFile",
  linux = TRUE)
}
\arguments{
\item{subject}{a object of class 'character'}

\item{text}{a object of class 'character'}

\item{address}{a object of class 'character'}

\item{file}{a object of class 'logical'}

\item{filename}{...}

\item{linux}{a object of class 'logical'}
}
\value{
a email is sent to you
}
\description{
Send a email upon R code completion
}

