% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_score.R
\name{z_score}
\alias{z_score}
\title{Calculate z-score for standardization}
\usage{
z_score(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a numeric vector standardized
}
\description{
Calculate z-score for standardization
}
\examples{
z_score(x = c(21, 15, 26, 26, 28, 29))
}

