% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_chr_to_factor.R
\name{convert_chr_to_factor}
\alias{convert_chr_to_factor}
\title{Convert character strings into factors}
\usage{
convert_chr_to_factor(df, ...)
}
\arguments{
\item{df}{a data frame}

\item{...}{dataframe with characters}
}
\value{
a new data.frame
}
\description{
a work in progress
}
\details{
http://stackoverflow.com/a/2851213
}
\examples{
 df = data.frame("H_member" = as.character(c("Jasmine", "Jenna", "Fiona")),
                           "current_age" = c(27, 27, 5))
convert_chr_to_factor(df)
}
